[WIP] Automatically use FIPS provider in FIPS mode
There are a couple of ways in which applications can be made to use the
FIPS provider:
- Editing openssl.cnf or loading applications with a custom configuration
(via the OPENSSL_CONF environment variable) in order to activate the FIPS
provider and specify the "fips=yes" property for algorithm fetches if
other providers are loaded. This doesn't work for applications that
override the default config or use a non-default context without manually
loading the default config.
- Programatically loading the FIPS provider and specifying the "fips=yes"
property for algorithm fetches if other providers are loaded, but this
requires code changes to individual applications.

This commit automatically makes use of the FIPS provider when a machine
is booted in FIPS mode by doing the following:
- Automatically sets "fips=yes" as the default property for all algorithm
fetches.
- Enables the FIPS provider as one of the predefined fallback providers
that are loaded if no providers are loaded via a configuration.

This approach works for all library contexts. If an application loads a
non-default provider config that disables the fallback provider loading,
then that config must load the FIPS provider so that algorithm fetches
don't fail. Applications can opt out of using the FIPS provider by
speciying the "-fips" or "fips=no" property for algorithm fetches.