crypto: Automatically use the FIPS provider when the kernel is booted in FIPS mode
This automatically configures all library contexts to use the FIPS provider when
the kernel is booted in FIPS mode by:
- Setting "fips=yes" as the default property for algorithm fetches
- Loading and activating the FIPS provider as the fallback provider.

If applications load providers via a configuration either because the default
configuration is modified or they override the default configuration, this
disables loading of the fallback providers. In this case, the configuration
must load the FIPS provider when FIPS mode is enabled, else algorithm fetches
will fail

Applications can choose to use non-FIPS approved algorithms by specifying the
"-fips" or "fips=no" property for algorithm fetches and loading the default
provider.