providers: Add a FIPS status indicator
FIPS 140-3 requires that services provide an indicator when the service uses an
approved cryptographic algorithm in an approved manner. In most cases, the
status can be determine implicitly by observing the return code from an
approved security service call, but the FIPS provider module may permit the
use of approved cryptographic algorithms in a non-approved manner in some cases.
For this, provide an explicit status indicator which is implemented as a
gettable provider module parameter, and can be accessed using
OSSL_PROVIDER_get_params() on the OSSL_PROVIDER associated with the FIPS provider
module.

The parameter is an integer which is 1 if the OSSL_LIB_CTX associated with the
FIPS provider's OSSL_PROVIDER has been used to execute a security service call
that used an approved cryptographic algorithm in a non-approved manner. Querying
the indicator automatically resets it to 0, as there's no mechanism to set
provider parameters.

Note that this only works for services inside of the FIPS module boundary. It
cannot be used to indicate whether the OSSL_LIB_CTX has executed security
service calls in other provider modules, and the ability to do this depends on
the provider configuration associated with the OSSL_LIB_CTX.