providers: Mark OSSL_FUNC_SIGNATURE_{SIGN,VERIFY} as unapproved
The OSSL_FUNC_SIGNATURE_SIGN and OSSL_FUNC_SIGNATURE_VERIFY functions sign
or verify a digest that was generated outside of the FIPS module bounary. As
FIPS 186-4 requires the use of an approved hash function, as defined in
FIPS 180, these functions cannot be an approved use of a signature service.