Version the module directory
The engines directory is already versioned so that the directory name changes
if the ABI breaks. In order to safely support provider modules that come from a
source other than the main openssl build, version the provider module directory
as well. This will enable us to split the openssl source package in Ubuntu (one
to provide libcrypto and the legacy module, and one to provide the FIPS module).