crypto/evp: add aliases for SHA2 digests
EVP_PKEY_CTX_get_ecdh_kdf_md relies on the namemap associated with the library
context to map an algorithm name to a legacy EVP_MD. This doesn't work reliably
if the EVP_KEYEXCH implementation associated with the EVP_PKEY_CTX lives inside
the FIPS provider because the application library context may not have fetched
any digests yet and may not contain a mapping for the digest algorithm name.
This breaks SHA2 digests because the primary names on the provider side don't
match any names that are added to application library context's namemap when it
is initialized.

This adds some aliases so that the primary names for SHA2 digests are added to
every application library context by default.