crypto/rsa: Tighten PSS salt length requirements in the FIPS provider
The OSSL_PKEY_RSA_PSS_SALT_LEN_AUTO_DIGEST_MAX option for RSA-PSS signing
selects a salt length that complies with FIPS 186-4 section 5.5.

Although this is the default option for the FIPS provider, it is still possible
to select a different option or salt length that doesn't comply with this
requirement. Add an extra check in the FIPS provider so that signatures with
salts that don't comply with FIPS 186-4 section 5.5 cannot be generated.

This also prevents verification of signatures with non-compliant salts, as FIPS
186-4 section 5.5 doesn't differentiate between signing and verification.