provider/fips: Move the definition of module_checksum into its own translation unit
An ealier change removed one of the two calls to verify_integrity. As it is now
only ever called with module_checksum as the expected checksum, and because
module_checksum is an array of zeroes at compile time (it is filled in post
processing), the compiler optimizes away the memcmp(expected, out, out_len) to a
simple check that the actual HMAC (out) is 32 bytes of zeroes:

   0x00007ffff7889eed <+653>:   mov    0xd8(%rsp),%rax
   0x00007ffff7889ef5 <+661>:   or     0xd0(%rsp),%rax
   0x00007ffff7889efd <+669>:   jne    0x7ffff7889f20 <verify_integrity+704>
   0x00007ffff7889eff <+671>:   mov    0xe8(%rsp),%rax
   0x00007ffff7889f07 <+679>:   or     0xe0(%rsp),%rax
   0x00007ffff7889f0f <+687>:   jne    0x7ffff7889f20 <verify_integrity+704>

Work around this by defining module_checksum in its own translation unit.